+++ tests/Router/Collector/RouteCollectorTest.php
@@ -7,8 +7,8 @@ namespace Membrane\OpenAPIRouter\Router\Collector;
 use cebe\openapi\Reader;
 use Membrane\OpenAPIRouter\Exception\CannotProcessOpenAPI;
 use Membrane\OpenAPIRouter\Exception\CannotRouteOpenAPI;
-use Membrane\OpenAPIRouter\Router\ValueObject\Route\Route;
-use Membrane\OpenAPIRouter\Router\ValueObject\RouteCollection;
+use Membrane\OpenAPIRouter\Router\Route\Route;
+use Membrane\OpenAPIRouter\Router\RouteCollection;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
